Fix issues, make consistent with main
