+++ data/interim/rfac_all_departing_and_arriving.csv
@@ -1,0 +1,4 @@
+all
+0
+0
+1
