+++ data/processed/validation_error_summary.csv
@@ -1,0 +1,13 @@
+validation_error_person,num_non_synthetic,num_synthetic
+"['general_use_transit_resident: Non-Critical', 'general_use_transit_resident: Non-Critical']",5,
+"['party_size_ground_access: Critical', 'reverse_mode: Non-Critical']",36,36
+"['party_size_ground_access: Critical', 'reverse_mode_predicted: Non-Critical', 'general_use_transit_visitor_home: Non-Critical', 'general_use_transit_visitor_home: Non-Critical']",1,
+"['party_size_ground_access: Critical', 'reverse_mode_predicted: Non-Critical']",6,
+"['reverse_mode: Non-Critical', 'general_use_transit_resident: Non-Critical', 'general_use_transit_resident: Non-Critical']",1,
+"['reverse_mode: Non-Critical', 'general_use_transit_visitor_home: Non-Critical', 'general_use_transit_visitor_home: Non-Critical']",41,41
+['reverse_mode: Non-Critical'],232,200
+"['reverse_mode_predicted: Non-Critical', 'general_use_transit_visitor_home: Non-Critical', 'general_use_transit_visitor_home: Non-Critical']",8,
+['reverse_mode_predicted: Non-Critical'],125,
+validation_error_trip,,
+main_mode,1,
+Total Complete Records,4731,4117
