+++ notebooks/02-survey-expansion.Rmd
@@ -23,14 +23,14 @@ print(getwd())
 
 # Remote I/O
 ```{r remote-io}
-setwd("/Users/wsp/Documents/GitHub/Survey_Airport/notebooks/")
+setwd(getwd())
 
 processed_dir <- "../data/processed/"
 interim_dir <- "../data/interim/"
 output_dir <- "../data/interim/"
 
-config_file <- paste0(interim_dir, "expansion_config_departing_only.yaml")
-# config_file <- paste0(interim_dir, "expansion_config_departing_and_arriving.yaml")
+#config_file <- paste0(interim_dir, "expansion_config_departing_only.yaml")
+config_file <- paste0(interim_dir, "expansion_config_departing_and_arriving.yaml")
 
 script_config <- yaml.load_file(config_file)
